renamed KeyboardScrollModel to ScrollModel made it so ScrollModel can be used by whatever child component it has
